requirements.txt
absl-py==0.9.0
AnyQt==0.0.10
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
BlackBoxAuditing==0.1.54
bleach==3.1.0
Bottleneck==1.3.1
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
cmake==3.16.3
commonmark==0.9.1
cryptography==2.8
cvxpy==1.1.0a3
cvxpylayers==0.1.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
diffcp==1.0.13
dill==0.3.1.1
docutils==0.16
easydict==1.9
ecos==2.0.7.post1
entrypoints==0.3
future==0.18.2
gast==0.3.3
glob2==0.7
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
jeepney==0.4.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
keyrings.alt==3.4.0
kiwisolver==1.1.0
lime==0.1.1.37
llvmlite==0.31.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mord==0.6
more-itertools==8.1.0
MulticoreTSNE==0.1
multiprocess==0.70.9
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==1.11
notebook==6.0.3
numba==0.42.0
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.32
opt-einsum==3.1.0
Orange3==3.7.1
osqp==0.6.1
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
py==1.8.1
pybind11==2.5.0
pycparser==2.19
Pygments==2.5.2
pyparsing==2.4.6
pyqtgraph==0.10.0
pyrsistent==0.15.7
pytest==5.3.3
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==18.1.1
requests==2.22.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
scs==2.1.0
seaborn==0.9.0
SecretStorage==3.1.2
Send2Trash==1.5.0
serverfiles==0.3.0
six==1.14.0
tensorboard==1.14.0
tensorboardX==2.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.0.0
torch==1.4.0
torchsummary==1.5.1
torchvision==0.5.0
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
tsnecuda==2.1.0
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
zipp==2.0.0
